requirements.txt
attrs==22.1.0
bcrypt==4.0.1
click==8.1.3
exceptiongroup==1.0.4
Flask==2.2.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-WTF==1.0.1
greenlet==2.0.1
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
meteostat==1.6.5
numpy==1.23.5
packaging==21.3
pandas==1.5.2
pluggy==1.0.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.10.0
